add write perf code
